Refactored code.

+++ core/src/main/java/com/tiktime/view/consts/ScreenConstants.java
@@ -6,8 +6,8 @@ public final class  ScreenConstants {
 
     public static final float PPM = 32;
 
-    public static final float VIEWPORT_WIDTH = VIRTUAL_WIDTH / PPM;
-    public static final float VIEWPORT_HEIGHT = VIRTUAL_HEIGHT / PPM;
+    public static final int VIEWPORT_WIDTH = (int) (VIRTUAL_WIDTH / PPM);
+    public static final int VIEWPORT_HEIGHT = (int) (VIRTUAL_HEIGHT / PPM);
 
     private ScreenConstants() {
         throw new UnsupportedOperationException("This is a utility class and cannot be instantiated");
